pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "bleach>=6.2.0",
    "fastapi>=0.116.1",
    "itsdangerous>=2.2.0",
    "jinja2>=3.1.6",
    "markdown>=3.8.2",
    "pydantic>=2.11.7",
    "pygments>=2.19.2",
    "python-dateutil>=2.9.0.post0",
    "python-multipart>=0.0.20",
    "pyyaml>=6.0.2",
    "starlette>=0.47.2",
    "uvicorn>=0.35.0",
]
